# server/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Market Info API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/stores/{store_id}/spotlight": {
      "get": {
        "tags": [
          "stores"
        ],
        "summary": "Get Store Spotlight",
        "operationId": "get_store_spotlight_api_v1_stores__store_id__spotlight_get",
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stores/{store_id}/products": {
      "get": {
        "tags": [
          "stores"
        ],
        "summary": "Get Store Products",
        "operationId": "get_store_products_api_v1_stores__store_id__products_get",
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/merchant/stories": {
      "post": {
        "tags": [
          "stories"
        ],
        "summary": "Create Story",
        "description": "LLM(또는 fallback)으로 스토리 생성 → Story 테이블 저장.",
        "operationId": "create_story_api_v1_merchant_stories_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "stories"
        ],
        "summary": "List My Stories",
        "operationId": "list_my_stories_api_v1_merchant_stories_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/merchant/stories/{story_id}": {
      "get": {
        "tags": [
          "stories"
        ],
        "summary": "Get My Story",
        "operationId": "get_my_story_api_v1_merchant_stories__story_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Story Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "stories"
        ],
        "summary": "Patch Story",
        "operationId": "patch_story_api_v1_merchant_stories__story_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Story Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "stories"
        ],
        "summary": "Delete My Story",
        "operationId": "delete_my_story_api_v1_merchant_stories__story_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Story Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/merchant/stories/{story_id}/publish": {
      "patch": {
        "tags": [
          "stories"
        ],
        "summary": "Patch Publish",
        "operationId": "patch_publish_api_v1_merchant_stories__story_id__publish_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Story Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryPublishRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stores/{store_id}/story": {
      "get": {
        "tags": [
          "stories"
        ],
        "summary": "Get Published Store Story",
        "description": "소비자용 — 점포에 게시된 최신 스토리. 없으면 null 반환.",
        "operationId": "get_published_store_story_api_v1_stores__store_id__story_get",
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "요청이 성공했습니다."
          },
          "data": {
            "title": "Data"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Meta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "BaseResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Meta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "timestamp"
        ],
        "title": "Meta"
      },
...
      "StoryPublishRequest": {
        "properties": {
          "publish": {
            "type": "boolean",
            "title": "Publish"
          }
        },
        "type": "object",
        "required": [
          "publish"
        ],
        "title": "StoryPublishRequest"
      },
      "StoryRequest": {
        "properties": {
          "store_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Id"
          },
          "save_to_store": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Save To Store",
            "default": false
          },
          "interview_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Text"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          },
          "tone": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "친근한",
                  "전문적인",
                  "정겨운"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Tone",
            "default": "친근한"
          },
          "selected_length": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "short",
                  "normal",
                  "detailed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Length",
            "default": "normal"
          },
          "publish": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publish",
            "default": false
          },
          "persist": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persist",
            "default": true
          }
        },
        "type": "object",
        "title": "StoryRequest"
      },
      "StoryUpdateRequest": {
        "properties": {
          "selected_length": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "short",
                  "normal",
                  "detailed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Length"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "title": "StoryUpdateRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}